#188 Added progressbar in splash screen activity
