+++ app/src/main/res/layout/splash_screen.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:background="@color/white">
+                android:layout_width="fill_parent"
+                android:layout_height="fill_parent"
+                android:background="@color/white">
 
     <ImageView
         android:id="@+id/imgLogo"
@@ -15,7 +15,20 @@
         android:background="@color/white"
         android:resizeMode="horizontal|vertical"
         android:scaleType="center"
-        android:src="@drawable/worship3" />
+        android:src="@drawable/worship3"/>
+
+    <ProgressBar
+        android:id="@+id/progress_bar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="67dp"
+        android:adjustViewBounds="true"
+        android:indeterminateTint="@color/light_navy_blue"
+        android:indeterminateTintMode="src_atop"
+        android:visibility="invisible"
+        android:resizeMode="horizontal|vertical"/>
 
     <!--<ProgressBar-->
     <!--android:id="@+id/progressBar1"-->
